(InitVariables): Add $RssInterwikiTranslate. (PrintFooter): Added admin bar inspired from UseMod 1.0. (ReBrowsePage): Third parameter is the anchor, UrlEncode all parameters. (DoPost): No longer encode id.
(BrowseResolvedPage): Add third parameter to call to ReBrowsePage.

(RSS): Use three-digit numbers to number items without a
timestamp, so that the number is not recognized as a valid
date (year-only) when printing the lines.

Support for Chinese pages, thanks to Ray and wctang.
(GetDownloadLink, BrowseResolvedPage, DoEdit): UrlEncode the page
id in the action.
(GetSearchLink): UrlEncode the name.
(DoPost): UrlEncode the page when calling ReBrowsePage..