Rewrote the code that skips over multiple pages.
When we're within a large rollback, $end used to point to the end of
the block that would eventually be stripped. It's a relative pointer.
While we were escanning downwards, we would sometimes strip single
pages even if we were within a larger rollback. Every time this
happened, $end was not shortened.

The current rewrite start skipping lines immediately and does away
with $skip_to and $end.