Anonymizing older entries in the list of recent changes.
When the maintenance action runs, it copies all the older entries in
rc.log to oldrc.log. Older entries are the entries that will usually
not get used by a typical display of recent changes. Any entry older
than the largest value of @RcDays is moved to oldrc.log (defaults to
90 days).

The idea is that you would only need hostnames or IP numbers to fight
spam and vandalism: Add regular expressions matching either hostname
or IP number of spammers or vandals to $BannedHosts and prevent the
attack from continuing. After a few days, however, this information is
no longer required. In this day and age of privacy invasion, I think
software should take a pro-active stance and therefore the entries
moved to oldrc.log will have their hostname or IP number replaced by
"Anonymous".

The existing entries in oldrc.log are not changed. If you want to do
the right thing, there's a script called anonymize.pl in the contrib
directory.