Get rid of $ReplaceForm
The old code decided whether the show the replacement field as part of
DoSearch and set $ReplaceForm. By that time, GetHeader had already
called GetSearchForm, so it was too late. I'm not sure why the
variable was necessary in the first place and so I'm removing it
entirely.

grep ReplaceForm *.pl modules/*.pl t/*.t comes up empty, now.