static-copy.pl: fix partial change from F to $F
The file writing operations had been only partially changed from F to
$F. Unfortunately, the old code relied on the fact that a file glob like
F was global in scope. To make the change to the file handle variable
more obvious, $F was renamed to $fh and this variable is now passed to
the subroutines StaticFile and StaticHtml.

In addition to that, :encoding(UTF-8) was changed back to :utf8 because
sometimes, this didn't work as intended.