Stop leaving locks behind
If you want to leave the thing locked when it is suddenly stopped (that is,
interrupting the process will lead to bad consequences, and the admin should
get in and make sure that everything is ok), then use **0**. **0** (or other
falsie) means do nothing if it is suddenly interrupted and leave the lock
in place.

**1** means that your lock will be removed automatically if the whole thing
is interrupted in the middle of the process. ##CleanLock('yourlock')##
function will be called, which means that you can add your own cleanup
logic by wrapping this subroutine (which I did in Modlue:private-wiki.pl).