Content Security Policy
For now, it is turned off by default, but we should try to turn it on as soon
as we can (this will require us to make sure that nothing in the codebase is
using inline javascript).

By default, everything from 'self' is allowed. This is good enough and I don't
think that it should be stricter.

style-src is set to * (anywhere!) because of the default stylesheet (which is
fetched from oddmuse.org) and also because of the “css=” query option.

img-src is set to * because linking to external images is a very common thing.