All CSP parameters should start with -
This is best described by the docs:
“Each argument name is preceded by a dash. Neither case nor order matters in
the argument list. -type, -Type, and -TYPE are all acceptable. In fact, only
the first argument needs to begin with a dash. If a dash is present in the
first argument, CGI.pm assumes dashes for the subsequent ones.”

We construct our own hash %headers and then pass that to $q->header. Since
hash values are not ordered in any way (in fact the order is pretty much
random), sometimes CSP things will be at the front of the list. Since there
was no leading dashes, these were interpreted as something else.